elasticsearch 의존성 추가 외...

+++ src/main/java/com/codeit/deokhugambatch/dashboard/popularbook/model/PopularBookCandidate.java
@@ -55,7 +55,7 @@ public class PopularBookCandidate {
 	/**
 	 * 평균 평점
 	 */
-	private double averageRating;
+	private BigDecimal averageRating;
 
 	/**
 	 * 좋아요 수
